// client/scripts/types/CryptoCompare.ts
 export interface USD {
    TYPE: string
    MARKET: string
    FROMSYMBOL: string
    TOSYMBOL: string
    FLAGS: string
    PRICE: number
    LASTUPDATE: number
    LASTVOLUME: number
    LASTVOLUMETO: number
    LASTTRADEID: string
    VOLUMEDAY: number
    VOLUMEDAYTO: number
    VOLUME24HOUR: number
    VOLUME24HOURTO: number
    OPENDAY: number
    HIGHDAY: number
    LOWDAY: number
    OPEN24HOUR: number
    HIGH24HOUR: number
    LOW24HOUR: number
    LASTMARKET: string
    VOLUMEHOUR: number
    VOLUMEHOURTO: number
    OPENHOUR: number
    HIGHHOUR: number
    LOWHOUR: number
    CHANGE24HOUR: number
    CHANGEPCT24HOUR: number
    CHANGEDAY: number
    CHANGEPCTDAY: number
    SUPPLY: number
    MKTCAP: number
    TOTALVOLUME24H: number
    TOTALVOLUME24HTO: number
    IMAGEURL: string
}

export interface EUR {
    TYPE: string
    MARKET: string
    FROMSYMBOL: string
    TOSYMBOL: string
    FLAGS: string
    PRICE: number
    LASTUPDATE: number
    LASTVOLUME: number
    LASTVOLUMETO: number
    LASTTRADEID: string
    VOLUMEDAY: number
    VOLUMEDAYTO: number
    VOLUME24HOUR: number
    VOLUME24HOURTO: number
    OPENDAY: number
    HIGHDAY: number
    LOWDAY: number
    OPEN24HOUR: number
    HIGH24HOUR: number
    LOW24HOUR: number
    LASTMARKET: string
    VOLUMEHOUR: number
    VOLUMEHOURTO: number
    OPENHOUR: number
    HIGHHOUR: number
    LOWHOUR: number
    CHANGE24HOUR: number
    CHANGEPCT24HOUR: number
    CHANGEDAY: number
    CHANGEPCTDAY: number
    SUPPLY: number
    MKTCAP: number
    TOTALVOLUME24H: number
    TOTALVOLUME24HTO: number
    IMAGEURL: string
}
// ...
export interface ShortCryptoPriceValue {
    USD: number
    EUR: number
}

export interface ShortCryptoPrice {
    [name: string]: ShortCryptoPriceValue
}